askme(NLP proj)/requirements.txt
streamlit
gemini
google-generativeai
docx
pyPDF2
io
re
zipfile
google.oauth2
googleapiclient.discovery
googleapiclient.http

